Pepper: Speed up destruction of QuotaReservation to prevent leaky tests.
Delete the QuotaReservation immediately when possible.

BUG=194304, 327759

Review URL: https://codereview.chromium.org/99333018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@240201 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5e818f92ea3507e1a219fdeddeadcb93fe0e48c2

+++ content/browser/renderer_host/pepper/quota_reservation.cc
@@ -116,12 +116,14 @@ void QuotaReservation::GotReservedQuota(
 }
 
 void QuotaReservation::DeleteOnCorrectThread() const {
-  if (file_system_context_) {
+  if (file_system_context_ &&
+      !file_system_context_->
+          default_file_task_runner()->RunsTasksOnCurrentThread()) {
     file_system_context_->default_file_task_runner()->DeleteSoon(
         FROM_HERE,
         this);
   } else {
-    // Unit testing code path.
+    // We're on the right thread to delete, or unit test.
     delete this;
   }
 }
